refactor(#108): v1, v2 패키지 분리
