+++ src/test/java/com/couponmoa/backend/domain/user/controller/v2/UserControllerV2Test.java
@@ -1,33 +1,34 @@
-package com.couponmoa.backend.domain.user.controller;
+package com.couponmoa.backend.domain.user.controller.v2;
 
 import com.couponmoa.backend.common.service.RedisService;
 import com.couponmoa.backend.config.JwtAuthenticationToken;
 import com.couponmoa.backend.config.JwtUtil;
 import com.couponmoa.backend.config.SecurityConfig;
 import com.couponmoa.backend.domain.user.dto.AuthUser;
+import com.couponmoa.backend.domain.user.dto.response.UserResponse;
+import com.couponmoa.backend.domain.user.entity.User;
 import com.couponmoa.backend.domain.user.enums.UserRole;
-import com.couponmoa.backend.domain.user.service.UserProfileService;
+import com.couponmoa.backend.domain.user.service.v2.UserServiceV2;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.context.annotation.Import;
-import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.util.ReflectionTestUtils;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.web.multipart.MultipartFile;
 
-import static org.mockito.ArgumentMatchers.any;
+import static org.hamcrest.Matchers.startsWith;
 import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.BDDMockito.willDoNothing;
+import static org.mockito.BDDMockito.given;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.authentication;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.multipart;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@WebMvcTest(UserProfileController.class)
+@WebMvcTest(UserControllerV2.class)
 @Import({SecurityConfig.class, JwtUtil.class})
-public class UserProfileControllerTest {
+public class UserControllerV2Test {
 
     @Autowired
     private MockMvc mockMvc;
@@ -36,7 +37,7 @@ public class UserProfileControllerTest {
     private RedisService redisService;
 
     @MockitoBean
-    private UserProfileService userProfileService;
+    private UserServiceV2 userServiceV2;
 
     private JwtAuthenticationToken userAuthenticationToken;
 
@@ -47,28 +48,21 @@ public class UserProfileControllerTest {
     }
 
     @Test
-    void 프로필_사진_등록() throws Exception {
-        MockMultipartFile file = new MockMultipartFile(
-                "image",
-                "profile.jpg",
-                "image/jpeg",
-                "fake image content".getBytes()
-        );
+    void 사용자_조회() throws Exception {
+        //given
+        long userId = 1L;
+        String email = "email@email.com";
+        User user = new User(email, "password", "name", UserRole.ROLE_USER);
+        ReflectionTestUtils.setField(user, "id", userId);
+        UserResponse mockResponse = UserResponse.fromEntityV2(user);
+        given(userServiceV2.findUser(anyLong())).willReturn(mockResponse);
 
-        willDoNothing().given(userProfileService).updateUserImage(anyLong(), any(MultipartFile.class));
-
-        mockMvc.perform(multipart("/api/v1/users/image")
-                        .file(file)
+        //when&then
+        mockMvc.perform(get("/api/v2/users")
                         .with(authentication(userAuthenticationToken)))
-                .andExpect(status().isOk());
-    }
-
-    @Test
-    void 프로필_사진_삭제() throws Exception {
-        willDoNothing().given(userProfileService).deleteUserImage(anyLong());
-
-        mockMvc.perform(delete("/api/v1/users/image")
-                        .with(authentication(userAuthenticationToken)))
-                .andExpect(status().isOk());
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.data.id").value(userId))
+                .andExpect(jsonPath("$.data.email").value(email))
+                .andExpect(jsonPath("$.data.imageUrl").value(startsWith("https://d2mm3xa8sjonwp.cloudfront.net")));
     }
 }
